+++ R/remove_data.r
@@ -9,6 +9,9 @@
 #' data in *data_object* that maps to this control batch is also removed.
 #' The dependency structure works as follows: *batch_keyword* / *batch_time* ->
 #' *data_control* -> *data_object* -> *data_score* -> *data_doi*.
+#' After using `remove_data`, run `vacuum_data` to free-up unused memory in
+#' the database file. Depening on the database size, `vacuum_data` might
+#' take some minutes for execution.
 #'
 #' @param table Database table from which the batch should be removed.  Object
 #' of type `character`.
@@ -39,6 +42,7 @@
 #'   control = 1,
 #'   object = 1
 #' )
+#' vacuum_data()
 #' }
 #'
 #' @export
@@ -47,6 +51,7 @@
 #' @importFrom dplyr filter
 #' @importFrom glue glue
 #' @importFrom rlang .data
+#' @rdname remove_data
 
 remove_data <- function(table, control = NULL, object = NULL) {
   .check_length(table, 1)
@@ -84,6 +89,15 @@ remove_data <- function(table, control = NULL, object = NULL) {
   } else {
     stop(glue("Error: 'table' must be either 'batch_keywords', 'batch_time', 'data_control', 'data_object', 'data_score', or 'data_doi'.\nYou provided {table}."))
   }
+}
+
+#' @export
+#' @importFrom DBI dbExecute
+#' @rdname remove_data
+
+vacuum_data <- function() {
+  out <- dbExecute(conn = gt.env$globaltrends_db, statement = "vacuum")
+  if (out == 0) message("Vacuum completed successfully.")
 }
 
 #' @title Remove from batch_keywords